+++ python/lsst/ts/mthexapod/mock_controller.py
@@ -41,6 +41,11 @@ from . import simple_hexapod
 # The real controller may use 0.15
 TRACK_TIMEOUT = 1
 
+# Model motor current as proportional to fractional velocity
+# and bus voltage as constant
+AMPS_PER_FRAC_SPEED = 1
+BUS_VOLTAGE = 100
+
 
 class MockMTHexapodController(hexrotcomm.BaseMockController):
     """Mock MTHexapod controller that talks over TCP/IP.
@@ -145,7 +150,6 @@ class MockMTHexapodController(hexrotcomm.BaseMockController):
             constants.MAX_ANGULAR_VEL_LIMIT,
         )
         # Order: x, y, z, u, w, v
-        config.initial_pos = (0, 0, 0, 0, 0, 0)
         config.pivot = self.pivot
         config.max_displacement_strut = self.actuator_max_length
         config.max_velocity_strut = self.actuator_speed
@@ -304,10 +308,21 @@ class MockMTHexapodController(hexrotcomm.BaseMockController):
             ]
             in_position = all(axes_in_position)
             self.telemetry.application_status = (
-                int(in_position) * ApplicationStatus.HEX_MOVE_COMPLETE_MASK
+                int(in_position) * ApplicationStatus.MOVE_COMPLETE
                 | ApplicationStatus.DDS_COMMAND_SOURCE
             )
             self.telemetry.input_pin_states = (0,) * 3
+
+            # TODO DM-31290: uncomment these lines when the data is available
+            # Model current and volage as proportional to fractional velocity
+            # axes_frac_velocity = [
+            #     actuator.velocity(tai=curr_tai) / actuator.speed
+            #     for actuator in self.hexapod.actuators
+            # ]
+            # self.telemetry.motor_current[:] = np.multiply(
+            #     axes_frac_velocity, AMPS_PER_FRAC_SPEED
+            # )
+            # self.telemetry.bus_voltage[:] = BUS_VOLTAGE
 
             # state, enabled_substate and offline_substate
             # are all set by set_state
@@ -321,20 +336,22 @@ class MockMTHexapodController(hexrotcomm.BaseMockController):
             self.telemetry.strut_encoder_raw = tuple(
                 pos * self.actuator_encoder_resolution for pos in current_lengths
             )
-            self.telemetry.strut_encoder_microns = tuple(current_lengths)
+            self.telemetry.strut_measured_pos_um = tuple(current_lengths)
 
             # self.telemetry.commanded_pos and commanded_length are both set
             # by MOVE and MOVE_LUT.
-            # self.telemetry.measured_pos should be based on current position,
-            # but SimpleHexapod does not yet support determining orientation
-            # from actuator length.
-            measured_pos = np.copy(self.telemetry.commanded_pos)
+            # self.telemetry.measured_xyz and uvw should be based on
+            # current position, but SimpleHexapod does not yet support
+            # determining orientation from actuator length.
+            measured_xyz = np.copy(self.telemetry.commanded_pos[:3])
+            measured_uvw = np.copy(self.telemetry.commanded_pos[3:])
             if self.telemetry.state == ControllerState.ENABLED:
                 # Add ~0.1 micron jitter to the current positions and
                 # ~0.003 arcsec jitter to the current rotations for realism.
-                measured_pos[:3] += self.xyz_jitter * (np.random.random(3) - 0.5)
-                measured_pos[3:] += self.uvw_jitter * (np.random.random(3) - 0.5)
-            self.telemetry.measured_pos = tuple(measured_pos)
+                measured_xyz += self.xyz_jitter * (np.random.random(3) - 0.5)
+                measured_uvw += self.uvw_jitter * (np.random.random(3) - 0.5)
+            self.telemetry.measured_xyz = tuple(measured_xyz)
+            self.telemetry.measured_uvw = tuple(measured_uvw)
             if (
                 self.telemetry.state == ControllerState.ENABLED
                 and self.telemetry.enabled_substate